chore: move some packges around so that we don't have bad imports
