+++ packages/core-components/src/layout/SignInPage/GuestUserIdentity.ts
